I've finally finished the port of the SmallBank benchmark in the OLTP-Bench framework. Yes, it took me three years but look I've been busy. I'm trying to run my trap at CMU. Shits been rough.
I would first like to give a shout out to KB. She keeps me out of trouble.

Second, I want to give a shout out to @eric-haibin-lin. He's been on point this semester with helping build our DBMS. There's no denying that.

Finally, and most importantly, shout out to @michaelcahill for sending me the original source code back in 2010 (or whenever it was). Michael is probably one of the best DB devs on the planet. Dope as fuck.

Ok that's it. The port is done. I've tested it on MySQL and Postgres. That's good enough for me. #41