Possible Bug Fix in AuctionMark (#245)
* Bug fix. getUniqueElementId(item_id, idx) incorrectly left shifts idx by 52 to merge it into item_id.
The least significant 56 bits of item_id are in use (see auctionmark/util/ItemId.java).
Replaced with left shift idx by 56.

* executeNewPurchase() (AuctionMarkWorker.java) was generating a random integer from 0 - 127
that is appended to the 56-bit ItemId for producing a Purchase Id. That was causing
duplicate primary key constraint exceptions when inserting into the ITEM_PURCHASE table.
Instead use a 7-bit per-ItemId counter with wrap-around.