Add Wikipedia set check to prevent adding the same watchlist page again.
I'm not sure exactly why this is an issue now.
Given the same (user id, page id), the rest is deterministic --
this was true before, this is true now.
It is possible that the issue simply never surfaced before.
In any case, this can be solved with a set for now.
A long-running Wikipedia benchmark which adds a lot of watchlist entries
may suffer in memory, but it looks like the default is biased heavily
towards not doing that.