Fix NoOp benchmark execution.
The benchmark always ran through the catch-block, because
stmt.executeQuery() expects a result. The single query ';' does not
return anything at all.

Now, the query is executed and if there is a result, than it is
processed as before.

Tested with PostgreSQL 13.

See also: https://github.com/oltpbenchmark/oltpbench/pull/229

---

This is a direct port of https://github.com/oltpbenchmark/oltpbench/pull/388
which was opened against OLTPBench after the work on BenchBase had begun.