Fixed running benchbase.jar from any directory
Previously, the current working directory had to be set to the same
location as the directory that contains benchbase.jar, as the code
expects a `config/plugins.xml` file to exist at the current directory.

This patch changes this so the jar and therefore the config/plugins.xml
location is automatically discovered. This allows benchbase to run from
any directory by doing `java -jar /path/to/benchbase.jar`.