Add configuration to read external ddl script
This commit adds a configuration option to the xml configuration files
that specifies a path to a ddl script. If the configuration option is
specified then that ddl script will be used instead of the default
script for the benchmark and database.

This could be useful for selecting between a row store and column store
layout or other alternative indexing options.

The xml tag is 'ddlpath'.

Resolves #153