containeruser UID/GID fixups
Use CONTAINERUSER_UID CONTAINERUSER_GID instead of UID/GID directly.

Allow those to be supplied as environment variable overrides.

If not available, default to UID/GID as before.

If those are 0 (root, which already exists in the container), then try
to look at the SUDO_UID environment variable.

Note: It doesn't yet handle the fixup required to the RUN command if
those are still set to 0.