CI checks on benchmark histograms and associated benchmark fixups (#268)
As previously noted in #122 the Wikipedia benchmark has a bug that results in most transactions getting aborted.

The fix developed in #173 by @apavlo seems to mostly fix it (though there are still some duplicate key errors generated by the benchmark's AddWatchPage procedure).

This PR pulls in those changes plus some additional ones (below) and additionally adds some CI checks on the json histogram outputs so that any benchmark the generates errors more than 1% of the completed transactions should be considered a failure (previously a CI failure was reported only if java exited non-zero, which doesn't happen when the many transactions fail, only when an unexpected exception occurs).

The 1% value is just an initial threshold and can be adjusted both on a per database and a per benchmark basis.
We do this in several places (e.g. TATP benchmark) in order to keep the changes required for this PR somewhat smaller and leave it for future work to improve those benchmarks.

To address the Wikipedia failures in some databases, we expand on auto-increment column support to SqlServer, add an auto-increment DDL schema for hsqldb for unit test purposes.

For others, we introduce improved schemas and/or dialects to fix query errors.

Note: in the case of sqlite, which currently lacks support for things like native `SLEEP` or `TABLOCK` we leave a more simplified version of the query for now and mark it for future work to improve.